This contains a large number of wholesale changes to the rest of ESBMC;
but very little ART related. The only actual change AFAIK was Lucas
importing the round robin code.

Conflicts:
	goto-symex/execution_state.h
	goto-symex/reachability_tree.cpp
	goto-symex/reachability_tree.h
	goto-symex/symex_main.cpp

+++ solvers/sat/pbs_dimacs_cnf.cpp
@@ -219,8 +219,8 @@ propt::resultt pbs_dimacs_cnft::prop_solve()
   pbfile.close();
 
   std::string msg=
-    i2string(no_variables())+" variables, "+
-    i2string(clauses.size())+" clauses";
+    i2string((unsigned long)no_variables())+" variables, "+
+    i2string((unsigned long)clauses.size())+" clauses";
   messaget::status(msg);
 
   bool result=pbs_solve();
